202504/0409/requirements.txt
requests==2.31.0
numpy==1.24.3
scipy==1.10.1
pandas==2.0.3
matplotlib==3.7.2
openpyxl==3.1.2
Pillow==10.0.0
tqdm==4.65.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
fake-useragent==1.3.0 